python/requirements.txt
numpy>=2.2.4
Pillow>=11.1.0
imageio-ffmpeg>=0.6.0
spandrel>=0.4.2,<0.5
packaging>=24,<27
